feat: added feature flag
